Refactor route handling and integrate PolylineService for improved navigation
- Replaced inline polyline generation logic with PolylineService methods across BookingManager, HomeScreen, and MapScreen for better code organization and maintainability.
- Enhanced route drawing functionality to utilize precomputed polylines, improving performance and user experience.
- Introduced fare calculation for routes based on polyline segments, streamlining fare updates during navigation.
- Updated camera movement utilities to ensure smooth transitions when displaying routes on the map.

+++ lib/utils/map_camera_utils.dart
@@ -6,7 +6,7 @@ Future<void> moveCameraToBounds(
   GoogleMapController controller,
   List<LatLng> coords, {
   double padding = 0.01,
-  int boundPadding = 20,
+  double boundPadding = 20.0,
 }) async {
   if (coords.isEmpty) return;
